Add CSS file to be used for all extensions/bookmarklet
This CSS file doesn't need to be maintained seperately - it would be
easiest to have it in a canonical place, and copy it elsewhere from
here.

Also, I've added `!important` to each CSS rule. This is kind of crazy..
I agree, but doing this, combined with eventually coming up with a
uniform naming scheme for our CSS attributes will get us much closer
to addressing PMT #103630.